Sonar cleanup.
